Refactor hybrid parser tests to own file
