+++ tests/test_common.py
@@ -49,306 +49,6 @@ def test_password(testdir):
     tables = camelot.read_pdf(filename, password="userpass", flavor="stream")
     assert_frame_equal(df, tables[0].df)
 
-
-def test_stream(testdir):
-    df = pd.DataFrame(data_stream)
-
-    filename = os.path.join(testdir, "health.pdf")
-    tables = camelot.read_pdf(filename, flavor="stream")
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_stream_table_rotated(testdir):
-    df = pd.DataFrame(data_stream_table_rotated)
-
-    filename = os.path.join(testdir, "clockwise_table_2.pdf")
-    tables = camelot.read_pdf(filename, flavor="stream")
-    assert_frame_equal(df, tables[0].df)
-
-    filename = os.path.join(testdir, "anticlockwise_table_2.pdf")
-    tables = camelot.read_pdf(filename, flavor="stream")
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_stream_two_tables(testdir):
-    df1 = pd.DataFrame(data_stream_two_tables_1)
-    df2 = pd.DataFrame(data_stream_two_tables_2)
-
-    filename = os.path.join(testdir, "tabula/12s0324.pdf")
-    tables = camelot.read_pdf(filename, flavor="stream")
-
-    assert len(tables) == 2
-    assert df1.equals(tables[0].df)
-    assert df2.equals(tables[1].df)
-
-
-def test_stream_table_regions(testdir):
-    df = pd.DataFrame(data_stream_table_areas)
-
-    filename = os.path.join(testdir, "tabula/us-007.pdf")
-    tables = camelot.read_pdf(
-        filename, flavor="stream", table_regions=["320,460,573,335"]
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_stream_table_areas(testdir):
-    df = pd.DataFrame(data_stream_table_areas)
-
-    filename = os.path.join(testdir, "tabula/us-007.pdf")
-    tables = camelot.read_pdf(
-        filename, flavor="stream", table_areas=["320,500,573,335"]
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_stream_columns(testdir):
-    df = pd.DataFrame(data_stream_columns)
-
-    filename = os.path.join(testdir, "mexican_towns.pdf")
-    tables = camelot.read_pdf(
-        filename, flavor="stream", columns=["67,180,230,425,475"], row_tol=10
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_stream_split_text(testdir):
-    df = pd.DataFrame(data_stream_split_text)
-
-    filename = os.path.join(testdir, "tabula/m27.pdf")
-    tables = camelot.read_pdf(
-        filename,
-        flavor="stream",
-        columns=["72,95,209,327,442,529,566,606,683"],
-        split_text=True,
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_stream_flag_size(testdir):
-    df = pd.DataFrame(data_stream_flag_size)
-
-    filename = os.path.join(testdir, "superscript.pdf")
-    tables = camelot.read_pdf(filename, flavor="stream", flag_size=True)
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_stream_strip_text(testdir):
-    df = pd.DataFrame(data_stream_strip_text)
-
-    filename = os.path.join(testdir, "detect_vertical_false.pdf")
-    tables = camelot.read_pdf(filename, flavor="stream", strip_text=" ,\n")
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_stream_edge_tol(testdir):
-    df = pd.DataFrame(data_stream_edge_tol)
-
-    filename = os.path.join(testdir, "edge_tol.pdf")
-    tables = camelot.read_pdf(filename, flavor="stream", edge_tol=500)
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_stream_layout_kwargs(testdir):
-    df = pd.DataFrame(data_stream_layout_kwargs)
-
-    filename = os.path.join(testdir, "detect_vertical_false.pdf")
-    tables = camelot.read_pdf(
-        filename, flavor="stream", layout_kwargs={"detect_vertical": False}
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid(testdir):
-    df = pd.DataFrame(data_stream)
-
-    filename = os.path.join(testdir, "health.pdf")
-    tables = camelot.read_pdf(filename, flavor="hybrid")
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid_table_rotated(testdir):
-    df = pd.DataFrame(data_stream_table_rotated)
-
-    filename = os.path.join(testdir, "clockwise_table_2.pdf")
-    tables = camelot.read_pdf(filename, flavor="hybrid")
-    assert_frame_equal(df, tables[0].df)
-
-    filename = os.path.join(testdir, "anticlockwise_table_2.pdf")
-    tables = camelot.read_pdf(filename, flavor="hybrid")
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid_two_tables(testdir):
-    df1 = pd.DataFrame(data_stream_two_tables_1)
-    df2 = pd.DataFrame(data_stream_two_tables_2)
-
-    filename = os.path.join(testdir, "tabula/12s0324.pdf")
-    tables = camelot.read_pdf(filename, flavor="hybrid")
-
-    assert len(tables) == 2
-    assert df1.equals(tables[0].df)
-    assert df2.equals(tables[1].df)
-
-
-def test_hybrid_table_regions(testdir):
-    df = pd.DataFrame(data_stream_table_areas)
-
-    filename = os.path.join(testdir, "tabula/us-007.pdf")
-    tables = camelot.read_pdf(
-        filename, flavor="hybrid", table_regions=["320,460,573,335"]
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid_table_areas(testdir):
-    df = pd.DataFrame(data_stream_table_areas)
-
-    filename = os.path.join(testdir, "tabula/us-007.pdf")
-    tables = camelot.read_pdf(
-        filename, flavor="hybrid", table_areas=["320,500,573,335"]
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid_columns(testdir):
-    df = pd.DataFrame(data_stream_columns)
-
-    filename = os.path.join(testdir, "mexican_towns.pdf")
-    tables = camelot.read_pdf(
-        filename, flavor="hybrid", columns=["67,180,230,425,475"], row_tol=10
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid_split_text(testdir):
-    df = pd.DataFrame(data_stream_split_text)
-
-    filename = os.path.join(testdir, "tabula/m27.pdf")
-    tables = camelot.read_pdf(
-        filename,
-        flavor="hybrid",
-        columns=["72,95,209,327,442,529,566,606,683"],
-        split_text=True,
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid_flag_size(testdir):
-    df = pd.DataFrame(data_stream_flag_size)
-
-    filename = os.path.join(testdir, "superscript.pdf")
-    tables = camelot.read_pdf(filename, flavor="hybrid", flag_size=True)
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid_strip_text(testdir):
-    df = pd.DataFrame(data_stream_strip_text)
-
-    filename = os.path.join(testdir, "detect_vertical_false.pdf")
-    tables = camelot.read_pdf(filename, flavor="hybrid", strip_text=" ,\n")
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid_edge_tol(testdir):
-    df = pd.DataFrame(data_stream_edge_tol)
-
-    filename = os.path.join(testdir, "edge_tol.pdf")
-    tables = camelot.read_pdf(filename, flavor="hybrid", edge_tol=500)
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_hybrid_layout_kwargs(testdir):
-    df = pd.DataFrame(data_stream_layout_kwargs)
-
-    filename = os.path.join(testdir, "detect_vertical_false.pdf")
-    tables = camelot.read_pdf(
-        filename, flavor="hybrid", layout_kwargs={"detect_vertical": False}
-    )
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_lattice(testdir):
-    df = pd.DataFrame(data_lattice)
-
-    filename = os.path.join(
-        testdir, "tabula/icdar2013-dataset/competition-dataset-us/us-030.pdf"
-    )
-    tables = camelot.read_pdf(filename, pages="2")
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_lattice_table_rotated(testdir):
-    df = pd.DataFrame(data_lattice_table_rotated)
-
-    filename = os.path.join(testdir, "clockwise_table_1.pdf")
-    tables = camelot.read_pdf(filename)
-    assert_frame_equal(df, tables[0].df)
-
-    filename = os.path.join(testdir, "anticlockwise_table_1.pdf")
-    tables = camelot.read_pdf(filename)
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_lattice_two_tables(testdir):
-    df1 = pd.DataFrame(data_lattice_two_tables_1)
-    df2 = pd.DataFrame(data_lattice_two_tables_2)
-
-    filename = os.path.join(testdir, "twotables_2.pdf")
-    tables = camelot.read_pdf(filename)
-    assert len(tables) == 2
-    assert df1.equals(tables[0].df)
-    assert df2.equals(tables[1].df)
-
-
-def test_lattice_table_regions(testdir):
-    df = pd.DataFrame(data_lattice_table_regions)
-
-    filename = os.path.join(testdir, "table_region.pdf")
-    tables = camelot.read_pdf(filename, table_regions=["170,370,560,270"])
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_lattice_table_areas(testdir):
-    df = pd.DataFrame(data_lattice_table_areas)
-
-    filename = os.path.join(testdir, "twotables_2.pdf")
-    tables = camelot.read_pdf(filename, table_areas=["80,693,535,448"])
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_lattice_process_background(testdir):
-    df = pd.DataFrame(data_lattice_process_background)
-
-    filename = os.path.join(testdir, "background_lines_1.pdf")
-    tables = camelot.read_pdf(filename, process_background=True)
-    assert_frame_equal(df, tables[1].df)
-
-
-def test_lattice_copy_text(testdir):
-    df = pd.DataFrame(data_lattice_copy_text)
-
-    filename = os.path.join(testdir, "row_span_1.pdf")
-    tables = camelot.read_pdf(filename, line_scale=60, copy_text="v")
-    assert_frame_equal(df, tables[0].df)
-
-
-def test_lattice_shift_text(testdir):
-    df_lt = pd.DataFrame(data_lattice_shift_text_left_top)
-    df_disable = pd.DataFrame(data_lattice_shift_text_disable)
-    df_rb = pd.DataFrame(data_lattice_shift_text_right_bottom)
-
-    filename = os.path.join(testdir, "column_span_2.pdf")
-    tables = camelot.read_pdf(filename, line_scale=40)
-    assert df_lt.equals(tables[0].df)
-
-    tables = camelot.read_pdf(filename, line_scale=40, shift_text=[""])
-    assert df_disable.equals(tables[0].df)
-
-    tables = camelot.read_pdf(filename, line_scale=40, shift_text=["r", "b"])
-    assert df_rb.equals(tables[0].df)
-
-
 @skip_pdftopng
 def test_repr_poppler(testdir):
     filename = os.path.join(testdir, "foo.pdf")
